% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_to_lonlat.R
\name{geom_to_lonlat}
\alias{geom_to_lonlat}
\title{Geographical Coordinates Converter}
\usage{
geom_to_lonlat(x)
}
\arguments{
\item{dataframe, }{only POINT-type}
}
\value{
dataframe columns long/lat
}
\description{
Splits WKT formatted coordinates into separate longitude/latitude columns
}
\keyword{WKT,}
\keyword{geography}
\keyword{geometry,}
\keyword{latitude,}
\keyword{longitude,}
