% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchColClasses.R
\name{matchColClasses}
\alias{matchColClasses}
\title{Dataframe Column Classes Match}
\usage{
matchColClasses(df1, df2)
}
\value{
dataframe
}
\description{
Mutates second dataframe columns to match the column classes of first dataframe,
returns second dataframe
}
\keyword{class,}
\keyword{classes,}
\keyword{dataframe}
\keyword{match,}
\keyword{mutate,}
